moved all header components out of CiimView
