+++ src/components/CiimView/CiimView.js
@@ -1,84 +1,25 @@
 import React, { useState } from 'react';
 import { 
-  Dropdown,
-  DropdownMenu,
   Pane,
-  PaneHeader,
   PaneMenu,
   Paneset,
 } from '@folio/stripes/components';
-import Divider from '../Divider';
 
 import {
   CollapseFilterPaneButton,
-  ExpandFilterPaneButton,
 } from '@folio/stripes/smart-components';
 import { FormattedMessage } from 'react-intl';
 
-import SearchBar from './HeaderComponents';
-import css from './CiimView.css';
+import { Header } from './HeaderComponents';
 
 export default function CiimView() {
   const [searchTerm, setSearchTerm] = useState('');
   const [showFilter, setShowFilter] = useState(true);
-  const [dropdownOpen, setDropDownOpen] = useState(false);
 
   const toggleFilterPane = () => {
     setShowFilter(!showFilter);
   };
-
-  const toggleDropdown = () => {
-    setDropDownOpen(!dropdownOpen);
-  };
-
-
-  const ToggleFilterPaneButton = () => (
-    !showFilter &&
-      <div className={css.filterPaneButton}>
-        <PaneMenu>
-          <ExpandFilterPaneButton
-            onClick={toggleFilterPane}
-          />
-        </PaneMenu>
-      </div>
-  );
-
-  const ActionButton = () => (
-    <div className={css.actionButtonContainer}>
-      <Dropdown
-        buttonProps={{
-          'marginBottom0': true,
-        }}
-        label={<FormattedMessage id="ui-ciim.ciimView.paneHeader.searchBar.actions" />}
-        open={dropdownOpen}
-        onToggle={toggleDropdown}
-      >
-        <DropdownMenu role="menu">
-          <span>This is our dropdown</span>
-        </DropdownMenu>
-      </Dropdown>
-    </div>
-  )
-
-  const Header = () => (
-    <PaneHeader
-      header={
-        <div className={css.paneHeaderContainer}>
-          <ToggleFilterPaneButton />
-          <SearchBar
-            onSearch={() => null}
-            setSearchTerm={setSearchTerm}
-            searchterm={searchTerm}
-          />
-          <div className={css.divider}>
-            <Divider />
-          </div>
-          <ActionButton />
-        </div>
-      }
-    />
-  );
-
+  
   return (
     <Paneset>
       {showFilter &&
@@ -86,7 +27,7 @@ export default function CiimView() {
           defaultWidth="20%"
           lastMenu={
             <PaneMenu>
-              <CollapseFilterPaneButton onClick={toggleFilterPane} />
+              <CollapseFilterPaneButton key="toggleFilterPane" onClick={toggleFilterPane} />
             </PaneMenu>
           }
           paneTitle={<FormattedMessage id="ui-ciim.ciimView.paneHeader.filter" />}
@@ -96,7 +37,14 @@ export default function CiimView() {
       }
       <Pane
         defaultWidth="fill"
-        renderHeader={Header}
+        renderHeader={() =>
+          <Header
+            searchTerm={searchTerm}
+            setSearchTerm={setSearchTerm}
+            showFilter={showFilter}
+            toggleFilterPane={toggleFilterPane}
+          />
+        }
       >
         <p> This is where the CIIM app will go </p>
       </Pane>
