+++ src/components/CiimView/HeaderComponents/ActionButton.js
@@ -1,0 +1,36 @@
+import React, { useState } from 'react';
+
+import { 
+  Dropdown,
+  DropdownMenu
+} from '@folio/stripes/components';
+
+import { FormattedMessage } from 'react-intl';
+
+import css from './Header.css';
+
+const ActionButton = () => {
+  const [dropdownOpen, setDropDownOpen] = useState(false);
+  const toggleDropdown = () => {
+    setDropDownOpen(!dropdownOpen);
+  };
+
+  return (
+    <div className={css.actionButtonContainer}>
+      <Dropdown
+        buttonProps={{
+          'marginBottom0': true,
+        }}
+        label={<FormattedMessage id="ui-ciim.ciimView.paneHeader.searchBar.actions" />}
+        open={dropdownOpen}
+        onToggle={toggleDropdown}
+      >
+        <DropdownMenu role="menu">
+          <span>This is our dropdown</span>
+        </DropdownMenu>
+      </Dropdown>
+    </div>
+  );
+}
+
+export default ActionButton;
