+++ src/components/CiimView/HeaderComponents/index.js
@@ -1,1 +1,7 @@
-export { default } from './SearchBar';
+
+// eslint-disable-next-line import/prefer-default-export
+export { default as SearchBar } from './SearchBar';
+export { default as Divider } from './Divider';
+export { default as ActionButton } from './ActionButton';
+export { default as ToggleFilterPaneButton } from './ToggleFilterPaneButton';
+export { default as Header } from './Header';
